refactor: rename CollapsibleGroup module to Collapsible
The module path `@xds/core/CollapsibleGroup` was unnecessarily specific —
the primary export is `XDSCollapsible`, with `XDSCollapsibleGroup` as the
coordination wrapper. Renaming to `@xds/core/Collapsible` better reflects
the module's purpose.

Changes:
- Rename `packages/core/src/CollapsibleGroup/` → `Collapsible/`
- Update package.json exports: `./CollapsibleGroup` → `./Collapsible`
- Update barrel export in `packages/core/src/index.ts`
- Update storybook import path
- Fix stale SYNC comments pointing to nonexistent stories file
- Update Card README cross-reference

No component or API changes — just the import path.
